refactor(encoders): split adapter grab-bag, top-level imports, dedupe tests
Split the 830-line hybrid_adapter.py into family files named for their
contents: token_adapters.py (full/global-token + global-embedding),
house_context_adapter.py, and house_points_adapter.py; hybrid_adapter.py
keeps only HybridObsAdapter. Pure move, no behavior change.

Convert the adapter lazy import_module() closures in the launcher classes
to normal top-level imports (no circular import; encoders already loads
JAX regardless). base.py's vggt_adapter import stays lazy — it is the one
that genuinely defers the heavy VGGT stack.

Dedupe test_encoders.py: replace ~14 near-identical inline FakeExtractor +
monkeypatch preambles with a shared _StubExtractor + patch_vggt fixture,
and fold four identical adapter-behavior fakes into two shared helpers
(1461 -> 1303 lines). Same 40 passed / 2 skipped.

Adds docs/notes/adr-encoder-adapter-simplification.md scoping the wider
registry refactor this is step 1 of.
