+++ src/r2dreamer/encoders/house_points_pose.py
@@ -3,12 +3,15 @@
 from __future__ import annotations
 
 from collections.abc import Mapping
-from importlib import import_module
 from types import MappingProxyType
 from typing import Any
 
 import flax.linen as nn
 
+from src.r2dreamer.adapters.house_points_adapter import (
+    VGGTHousePointsPoseObsAdapter,
+    VGGTHybridHousePointsPoseObsAdapter,
+)
 from src.r2dreamer.encoders.base import VGGTEncoder
 from src.r2dreamer.encoders.mlp import (
     HousePointsCameraEncoder as ModelHousePointsCameraEncoder,
@@ -78,8 +81,7 @@ class VGGTHousePointsPoseEncoder(VGGTEncoder):
         return True
 
     def _build_adapter_for_extractor(self, extractor):
-        adapter_module = import_module("src.r2dreamer.adapters.hybrid_adapter")
-        return adapter_module.VGGTHousePointsPoseObsAdapter(
+        return VGGTHousePointsPoseObsAdapter(
             extractor,
             house_points_path=self._house_points_path,
         )
@@ -113,8 +115,7 @@ class VGGTHybridHousePointsPoseEncoder(VGGTHousePointsPoseEncoder):
         )
 
     def _build_adapter_for_extractor(self, extractor):
-        adapter_module = import_module("src.r2dreamer.adapters.hybrid_adapter")
-        return adapter_module.VGGTHybridHousePointsPoseObsAdapter(
+        return VGGTHybridHousePointsPoseObsAdapter(
             extractor,
             house_points_path=self._house_points_path,
         )
